time fix

+++ src/metatrader_position/metatrader_position.rs
@@ -50,6 +50,6 @@ pub struct MetatraderPositionBodySbModel {
     #[prost(enumeration = "PositionEventType", tag = "8")]
     pub event_type: i32,
 
-    #[prost(int32, tag = "9")]
-    pub time: i32,
+    #[prost(int64, tag = "9")]
+    pub time: i64,
 }
